+++ src/main/java/lotto/WinningStandard.java
@@ -7,19 +7,19 @@ public enum WinningStandard {
     SECONDPLACE(5, true, 30000000, "30,000,000"),
     FIRSTPLACE(6, false, 2000000000, "2,000,000,000");
 
-    private final int matchingNumber;
+    private final long matchingNumber;
     private final boolean matchingBonusNumber;
-    private final int prize;
+    private final long prize;
     private final String prizeForPrint;
 
-    WinningStandard(int matchingNumber, boolean matchingBonusNumber, int prize, String prizeForPrint) {
+    WinningStandard(long matchingNumber, boolean matchingBonusNumber, long prize, String prizeForPrint) {
         this.matchingNumber = matchingNumber;
         this.matchingBonusNumber = matchingBonusNumber;
         this.prize = prize;
         this.prizeForPrint = prizeForPrint;
     }
 
-    public int getMatchingNumber() {
+    public long getMatchingNumber() {
         return matchingNumber;
     }
 
@@ -27,7 +27,7 @@ public enum WinningStandard {
         return matchingBonusNumber;
     }
 
-    public int getPrize() {
+    public long getPrize() {
         return prize;
     }
